+++ diversitree/src/GslOdeTime.cpp
@@ -1,7 +1,7 @@
 #include "GslOdeTime.h"
 #include <R.h>
 
-GslOdeTime::GslOdeTime(SEXP extPtr, int size, TimeMachine tm_) :
+GslOdeTime::GslOdeTime(SEXP extPtr, size_t size, TimeMachine tm_) :
   GslOdeBase(size), tm(tm_) {
   // This generates a warning that we may live with according to BDR:
   // https://stat.ethz.ch/pipermail/r-devel/2004-September/030792.html
@@ -20,6 +20,6 @@ void GslOdeTime::derivs(double t, const double y[],
   // TODO: An iterator approach might be better than the extra copy
   // that this involves?
   std::vector<double> pars = tm.get(t);
-  derivs_f(size(), t, &pars[0], y, dydt);
+  derivs_f(static_cast<int>(size()), t, &pars[0], y, dydt);
 }
